Add function to set mysql env vars. PYTHON-1227
mysql_settings() is a new function that works in the same way as the
existing postgresql_settings() function. It's purpose is to centralize
the database configuration in one place for all of the mysql tests.

It will return the correct configuration depending on which environment
it is running in: Local, Tddium, or in the test docker container.