Remove quay.io dependency for squid docker image. PYTHON-1228
Try to pull the squid docker image from quay.io, in order to avoid
the lengthy build process. But, if 'docker pull' fails, ignore the
error and build it the slow way.

Previously, there was no fallback, so if the 'docker pull' command
failed, the script would throw an error, rather than building the
squid image from the Dockerfile.