Add wheels to devpi-server docker image. PYTHON-1258
Add build_wheels.sh script and call it from the Dockerfile. This script will
build a wheel for each package listed in the package-lists/wheels-* files.

The end result is that when tox creates a virtualenv and installs packages,
it is using as many cached wheels as possible, so that setting up the
test environment is as fast as possible.