Modify create_docker_base_images() to accept args. PYTHON-1266
This function in packnsend now accepts space-separated image names, so it's
possible to build a single image, or a subset of images.

The default behavior remains unchanged. If no images are passed on the
command line, then create_docker_base_images() will build all images listed
in the ALL_IMAGES variable.