Modify stop_application_services() to accept args. PYTHON-1266
This function in packnsend now accepts space-separated image names, so it's
possible to stop a single container, or a subset of containers.

The default behavior remains unchanged. If no images are passed on the
command line, then stop_application_services() will stop all packnsend
containers. (One container for each of the images listed in RUNNABLE_IMAGES.)