Add process_host.display_name setting. PYTHON-1241
By default, process_host.display_name is set to the value of
socket.gethostname(). But, it's possible to override this value
by setting it in newrelic.ini.

As per the spec, it's also possible to set process_host.display_name
using an environment variable: NEW_RELIC_PROCESS_HOST_DISPLAY_NAME.