Make sure 'agent_config' exists in collector response. PYTHON-1240
The response the collector sends for a 'connect' response has a different
format, depending on whether the customer has activated server-side
configuration or not. If it has been activated, then the 'agent_config'
dict will be present in the response. Otherwise, it won't.

We need to handle both cases, since it's possible to activate High
Security Mode without activating server-side configuration.