Fix tests, now that we send background tasks to Insights. PYTHON-1290
In `agent_features/test_analytics.py`, one test broke because the test said
a background task should not be recorded to Insights, when now it should be.

Five tests broke because we are no longer recording 'webDuration' in
StatsEngine.record_transaction(). That's correct behavior, according to the
latest version of the Agent Attributes spec. To fix the tests, I modified
`validate_analytics_sample_data()` so that it no longer checks for the
presence of `record['webDuration']`.