Instrument urlretrieve to handle file URI. PYTHON-1220
If the URI is a 'file:///' style URI, perform a FunctionTrace for
urllib.urlretrieve, so we don't display a host of 'unknown' in APM UI.
If the URI is not a 'file:///' style URI, continue to treat it as an
external trace.

Also, use the newer instrumentation API from newrelic.agent.