Modify urllib3 instrumentation to capture host. PYTHON-1304
Previously, we instrumented urllib3.request.RequestMethods.request_encode_url
and request_encode_body, but that only gives us access to the url, which is
often in the form of a path ('/index.html').

Instead, instrument a lower-level method, HTTPConnectionPool._make_request,
which gives us access to the host, so we can construct a complete url, which is
what our ExternalTrace expects.

Also, update instrumentation to the newrelic.agent API, by using ExternalTrace
and wrap_function_wrapper.