Add Synthetics resource id to TT. PYTHON-1376
The resource id should be included in the "Intrinsics" section, but
since we don't support the Agent Attributes spec yet, we include it
as a custom attribute.

The attribute name is 'nr.synthetics_resource_id' because the 'nr.'
prefix hides it in the APM UI, and keeps it from showing in the
"Custom attributes" section of a transaction trace.