Move synthetics ids to 'record' dict in analytic event. PYTHON-1376
In practice, it doesn't really matter whether the synthetics ids are
stored as custom parameters, or in the 'record' dict, since they all get
mushed together when they hit Dirac. But, it's more accurate to store
them in 'record', because they are added by the agent, not added by
the customer.