Refactor override_application_settings fixture. PYTHON-1405
This fixture is useful for overriding application-wide settings,
but with the new extra_settings function being used, some instrumentation
functions are looking for settings in different settings objects.
For example, in framework_django.py, a `django_settings` object is
created to hold the configuration for whitelisting inclusion tags.

For testing, it's necessary to override these framework-specific
settings, hence the refactor:

    Old: override_application_settings(settings)

    New: override_settings(settings, settings_object=None)

override_application_settings still exists, so I haven't changed all of
the tests that use it. But, if we're happy with the new override_settings
fixture, then we should go back and refactor all the tests that use
override_application_settings.