Refactor how to get Inclusion tag filename. PYTHON-1405
Refactor to make a little more robust. For instance, avoid KeyErrors and
IndexErrors when getting node_class by checking cells first. And, make
sure the name of the function we care about ('generic_tag_compiler') is
in the list of frame.f_code.co_names when we get the filename.