Increase timeout for EC2 metadata API calls. PYTHON-1662
I'm concerned a timeout of 0.25 is too short, so I'm bumping it to
0.5 seconds.

Other agents are setting a timeout of 1 second for all 3 calls to the
metadata API, while we are setting individual connect and read timeouts
for each call.

Originally, I chose 0.25 seconds as the timeout, because that would result
in a worst case scenario of a delay of 1.5 seconds. (Each of the 3 calls
would return successfully just before the timeout.) While that's roughly
in line with the worst-case scenario for other agents, that doesn't take
into account the fact that our worst-case scenario is much less likely than
other agents.

When a user's application is not running on EC2, our first API call fails
with a read timeout, so the more typical case for us is a delay of 0.25
seconds.

Because we stop querying after the first failure, we can afford a longer
timeout per call. With this change, our delay for non-EC2 applications
will be 0.5 seconds, which is still shorter than most agents. And, our
worst-case scenario of 3 seconds of delay is not only highly unlikely,
but it's also not too horrible, if it does happen.