Rough first pass instrumenting stack_context.py PYTHON-1691
This is still a work in progress but I wanted to push it up so others could give
it a look since I'll be out next week. Issues have been highlighted with TODOs
in the code. One in particular is that we may need to end the transaction later
than on a call to RequestHandler.on_finish() since that is added as a callback
on the current transaction (so it won't be complete if we exit the transaction
in on_finish).