Require settings arg when creating AttributeFilter. PYTHON-1701
Explicitly passing it in is better, since we are less likely to use the
original global settings by mistake, but we should always use the settings
that have had server-side config applied.

Plus, this allows the removal of the import, which solves a circular
import problem I'm having when moving the attribute filter from the
application object to the settings object.