Move attribute filter to application settings. PYTHON-1701
Given that we need access to the attribute filter in various places
(like StatsEngine, or TransactionNode), it's easier to get at it if it
is stored in the application settings object, rather than stored as an
attribute on the application object. That's because the application
settings object is also stored as an attribute of StatsEngine,
Transaction, and TransactionNode.

Note that the application settings object is different than the global
settings object, since server-side settings have been applied and the
attribute filter has been created. Both of those modifications to the
settings object take place in a new function:

    newrelic.core.config:finalize_application_settings