Be more careful with capture_params tri-state setting. PYTHON-1684
The default value for 'capture_params' is now None, so it's no longer
good enough to just check if it's True, and flip it to False if it is.
We need to be more explicit how we check for for True/False/None, in
both the function, as well as in tests.