Don't capture request params in High Security Mode. PYTHON-1684
Up to now, our strategy has been to capture request parameters always,
and then, if we are in High Security Mode, filter them out before
sending parameters to the collector.

However, if we aren't ever going to need request parameters in
High Security Mode, then there's no reason to capture them in the
first place.

Consider this defense in depth. We won't capture them. But, even
if we did somehow, we will still filter them out.