Prevent bypass of high security check. PYTHON-1684
Make add_customer_parameters() call add_custom_parameter(), so
that we can enforce the check for High Security Mode in one place
for adding custom parameters.

Note that `add_custom_parameters()` is not part of the public API
for the agent, so it's highly unlikely that customers were using
this.