Properly clear transaction from cache when it ends. PYTHON-1703
In transaction aware functions we do the following:

1) Replace the transaction in the transaction cache with the transaction
associated with the function. We save the old transaction.
2) We call the transaction aware function.
3) We replace the transaction in the transaction cache with the old
transaction.

There is a problem with this strategy in the case where we finalize the
transaction and the old transaction is identical to the transaction associated
with the transaction aware function. In step 3 we then put a transaction in the
transaction cache which has already completed.

This PR changes the behavior so we don't place a transaction into the
transaction cache after it ends.