Improve browser ASCII comment. PYTHON-1652
The reason we attempt to encode the browser monitoring content
is because we don't want to screw up the web page we are
inserting into. For example, if there is not meta tag for
encoding, and the browser encounters a character in our injected
content it doesn't know what to do with, it can make the wrong
assumption about the page encoding.