Default value of capture_params should be None. PYTHON-1697
For all web transactions, the default value of capture_params will get
overridden with the configuration setting in WebTransaction.__init__().
And, capture_params doesn't affect non-web transactions (background tasks
don't have request params).

So, this should have no real effect.

But, even so, it's better for the default value to be correct.