Add host.displayName agent attribute. PYTHON-1646
If the `process_host.display_name` configuration setting is set,
either in the newrelic.ini config file or with the environment
variable, then we add `host.displayName` to the agent attributes.

The spec says "Omit the attribute if the Display Host is the same
as the Host." However, there is an edge case where our implementation
will send it anyway, even if they match. We will send the attribute
if the user explicitly sets `process_host.display_name` to the same
value as the hostname.

This edge case will rarely happen, since it's extremely unlikely that
a customer will set the custom display name to be the same as the hostname
returned by socket.gethostname().

We could avoid that case by calling socket.gethostname() inside of
Transaction.agent_attributes and comparing, but I'd rather not pay the
performance penalty of calling gethostname() every time just to avoid
such a rare possibility.

Plus, if the customer sets `process_host.display_name` to a custom display
name, I think it should be added as an agent attribute, no matter what it's
value is.