Instrument exceptions with no ExceptionStackContext. PYTHON-1755
If a callback is added to the IOLoop from a function not running on the IOLoop, no StackContext is created (since it is the top of the stack). If an exception occurs in this function our instrumention of ExceptionStackContext will not catch it. This code change catches exceptions occuring in this case.