Add deploy-to-download.sh script. PYTHON-1761
By default, deploy to the `testing` subdirectory on the download
servers. Set the DOWNLOAD_DIR environment variable, if you want
to upload to the `release` subdirectory.

Use `rsync` rather than `scp` in order to ensure that this script
will not overwrite existing files. Rsync has an `--ignore-existing`
option, but `scp` lacks similar functionality.