Add unit tests for StatsEngine.custom_events. PYTHON-1792
We don't currently have any unit tests specifically for StatsEngine.
At some point, maybe we can add a complete set of tests?

But, for now, I've added tests that just focus on custom events. They
verify that the new custom_events attribute/property is initialized
properly, and that it is also reset with the correct capacity.