Record custom event to Transaction object. PYTHON-1794
Even though custom events are not associated with a transaction,
it makes sense to store them on the Transaction object so that
we can defer the work of adding them to the application's StatsEngine
object until after the response has been returned.

This fits our general strategy of limiting the amount of work we
do during the course of a transaction as much as possible.

The alternative would be to lock the application's StatsEngine object
and add it during the course of the transaction. We will have to do
this when adding a custom event outside of a transaction (similar to
how we handle recording exceptions), but when we can avoid doing that,
we should.