Update rum_client_config.json. PYTHON-1811
This update adds a test for when the collector sets js_agent_file
to an empty string in the response to connect(). Mhayes's pull
request (cross_agent_tests/pull/62) is the source for this test.
It hasn't been merged into master yet, but staging collector is
already sending an empty string.