Explicitly mark package as not zip safe. PYTHON-1815
Since finding the location of the CA certificates bundle is dependent
on it existing in the filesystem, we must make sure that the package is
never installed as a zip file.

Up until now, when installing the newrelic package as an egg file
(either from `easy_install newrelic`, or from `python setup.py bdist_egg
&& easy_install <newrelic egg>`), the package would be unzipped during
installation, because the presence of c extensions and the usage of
`__file__` would mark the package as not-zip-safe.

That means the newrelic package has never been installable as a zip
file.

So, adding the `zip_safe = False` flag to setup.py is just extra insurance
It results in the same outcome, and, if ever installed with easy_install,
there's a performance bonus, since it also skips the need to analyze the
package to determine if it is zip safe or not during install.