Add custom events directly to Application SampledDataSet. PYTHON-1793
In order to avoid the possibility of double-sampling custom events
when merging a single transaction into the Application's StatsEngine,
we add custom events directly from the TransactionNode to the
Application's SampledDataSet in Application.record_transaction().

Note that this is different from how we handle other events (Transaction
Events, Error Events, etc.), because for those types of events, we never
capture more events during the course of a single transaction than can fit
in the capacity of a SampledDataSet. Therefore, no double-sampling takes
place.

For custom events, it's possible to capture more custom events than the
SampledDataSet can hold during a single transaction, so we have to worry
about double-sampling.

Once we can change how we merge SampledDataSets and preserve the relative
weighting, so that all events have an equal chance of being included in
the merged SampledDataSet, then we can go back to treating all events the
same way.

Given the change in how we add custom events, we also need to change the
test fixtures:

* Removed:
    validate_custom_event_inside_transaction
    validate_transaction_record_custom_event

* Renamed:
    validate_custom_event_outside_transaction
    to
    validate_custom_event_in_application_stats_engine