Check config before creating a custom event. PYTHON-1843
To test that the two record_custom_event methods short-circuit, if the
custom events feature is disabled, I added yet another test fixture:

    function_not_called

On the one hand, this duplicates basic Mock functionality, and adds to
our fixture explosion. On the other hand, it avoids the need to add
another dependency on Mock.