Add flag to gate transaction closing. PYTHON-1875
After a transaction is created some methods are run synchronously. Within these
methods a callback can be scheduled and then waited on to complete. This can
make the refcount go to 0 before the request handler code is called. This PR
prevents the transaction from closing until the request is handled. That is for
a 'headers_received' called, which initiates the transaction, we prevent the
transaction closing until 1 of 2 methods is called, 'finished' or
'on_connection_closed'. One but not both of these are guaranteed to be called
according to the tornado documentation.