Protect against Futures resolving in a thread. PYTHON-1881
Although Tornado futures are explicitly described as not thread-safe,
We want to prevent any activity that happens as a result of a future
resolving in a thread to affect the main transaction. This is in line
with the way we (don't) support threads currently, and prevents the
agent from taking down an app that may have worked without New Relic
installed.