Enable background tasks to run on tornado IOLoop. PYTHON-1862
For Tornado4 web transactions, we set additional attributes on the
Transaction object, so that we can finalize the transaction after
all of it's callbacks have run, even if they occur after the web
response has been sent back.

Now, we can have similar functionality for background tasks. When a
function that's wrapped as a background task adds callbacks, the
transaction will end after the callbacks run, not after the function
exits.