Switching order of stack_context.wrap and newrelic wrapper. PYTHON-1871
Previously, we intercepted stack_context.wrap so when it went to wrap a
function, f, we would first apply our new relic wrapper to f and then apply
stack_context.wrap. The issue with this strategy is that stack_context.wrap
catches and handles exceptions occurring in f. We want to associate these
exceptions with the transactions they occur in. This seemed to work previously
because, when an exception occured, we wouldn't properly remove the exception
from the transaction cache. This leaves a transaction improperly in the cache.
Using context managers for transaction swapping ensures we don't improperly
leave a transaction in the cache but has uncovered this behavior.