Add test to end a transaction via remove_timeout without threads. PYTHON-1889
We currently test that a transaction can end via ioloop.remove_timeout by
using threads. We use threads because we want to end the transaction in a
function that is not associated with the transaction (otherwise this function
keeps the transaction open). However, remove_timeout is not thread safe. This
PR modifies this test to instead schedule a callback on the ioloop. The callback
is not associated with the transaction because we manually wrap it in a None
transaction context.