Avoid double wrapping function traces in tornado4. PYTHON-1887
I change create_transaction_aware_fxn to take a parameter that controls whether
we want to trace the fxn. There are cirumstances where we want to make a
function aware but do not want to trace it. For example, if a function is
decorated with @function_trace this function will be traced but it will not
yet be associated with a transaction. When we add this function as a callback we
will want to associate it with a transaction but we don't want to trace it since
it is already wrapped in a function tracer.