Move data source registration under instance lock. PYTHON-1521
Since we are using Agent._instace as a check to whether data sources
have been registered, in the remove utilzation code, we must have
a guarantee that the data source registration has then also run.
Otherwise, we could attempt to remove the thread utilization data
source *before* is actually registered - and miss our chance to
remove it.