Increment/decrement ref_count in Runner. PYTHON-1905
Prior to this fix, yielding from a coroutine could cause a transaction
to end prematurely, causing us to fail to include work that belonged
in the transaction, and also causing us to log an error.

Now, we ensure that yielding from a coroutine must complete before
we end the transaction. We do this by incrementing the ref_count in
Runner.__init__ and decrementing it when the Runner finishes in
Runner.run().