Disable explain plans for queries made in async mode. PYTHON-1965
If we don't save the connection parameters for a query, then we don't
run an explain plan for it. So, we add a check to make sure the
database connection was not in asynchronous mode before adding the
connect_params.

Note that only the psycopg2 driver uses the `async` keyword argument.