Packnsend docker tests. Python-1983
Creates the multijob for the packnsend docker tests. This is a conversion of
the Python_Agent-TESTS-DOCKER-NEW job from pdx-hudson. The jobs are configured
in a json file which sets the job phase, name, description, and commands.

Prior to each test run, the job will do a `packnsend pull_with_restart`. This
will pull the packnsend docker images from dogestry and restart any that have
changed. This means that upgrading the packnsend images is as easy as pushing
them to dogestry.

Test can be run on command and every hour with cron.

Jobs take about 18 min to run.