Ignore any utilization configs set server side
This is done by removing those items from the server_settings object that was
received from the server. Later, these values are then merged with the settings
found locally.

This implementation is consistent with the handling of high_security mode
settings found on the server. If found, any high_security related settings are
altered in the server_settings object to prior to its merging with the local
settings.