Simplify autorum disablement, add more tests. PYTHON-2006
Responding to feedback on PR
https://source.datanerd.us/python-agent/python_agent/pull/309

Instead of checking if the transaction is a websocket before applying the
middleware, simplify the logic by wrapping the decision up in the
autorum_disabled boolean.

Note on why we want to disable RUM for websockets:

Due to a bug in gevent-websocket (0.9.5) package, if our
_WSGIApplicationMiddleware is applied a websocket connection cannot be made.
The gevent-websocket package incorrectly handles applications that return
generators (which our middleware is a type of), therefore middleware
application is avoided. It doesn't make sense for websockets to include RUM
anyway.