Pass custom wrapped _start_response to wsgi application. PYTHON-2007
Fixes issue where the WebTransaction.status_code is not set when the customer
has disabled browser monitoring. (It is probably true that other attributes
were also not set)

The `process_response` method does several things including gathering info on
the response code and collecting headers. This method is called from
_start_response. However, in commit
https://source.datanerd.us/python-agent/python_agent/commit/1fa8e551f157a3fa35ef00c1d13be8c13c412097,
the arguments sent to `wrapper` (the wsgi application object) changed,
replacing `_start_response`. Undoing this change fixes the issue.

This change is the correct one and will (hopefully) not break anything because
one can observe that the line `result = wrapped(environ, _start_response)` was
used clear back to the beginning of time. For a specific example, when the
middleware layer was added in
https://source.datanerd.us/python-agent/python_agent/commit/81eccead4a024add65ccbd78cd4ee927a1c5f4ae
the line `result = wrapped(environ, _start_response)` was not changed. It was
moved.

It is unclear to me why `_start_response` was removed. The commit messages and
Jira tickets (https://newrelic.atlassian.net/browse/PYTHON-1381) offer no
guidance.