Fix error when Celery task runs during inactive transaction. PYTHON-2009
If a transaction is made inactive (stopped or ignored), and then a Celery
task is run, then the agent throws a RuntimeError.

Why? Because the agent sees no active transaction when it runs the task,
so it attempts to create a new Background transaction and save it to the
TransactionCache. However, the existing transaction already is in the
TransactionCache.

The problem is that the agent did not check if it was in an *inactive*
transaction. It only handled the cases where the task was run in an active
transaction, or if it ran outside of a transaction. With this commit, we now
address the "inactive Transaction" case in our Task instrumentation.

One way to cause this error is to create a "Celery Chord", which is composed
of multiple Celery tasks, and then call `newrelic.agent.ignore_transaction()`
during the middle of a "Chord." When the next task runs after
`ignore_transaction()`, then the agent throws the RuntimeError.

Our tests do not specifically create a Celery Chord, because that requires
running Celery tasks asynchronously, which would be a much more difficult
test to set up. However, we can reproduce this error scenario by making
synchronous calls to the Task functions and the agent API functions.