Link to original deploy jobs on PDX Jenkins. PYTHON-2022
Using the original jobs is important to keep the build numbers
aligned with previously released agents. (Build numbers should
monotonically increase, because they are part of the agent version
number, but running an entirely new deploy job will have a build
number of 1.)

These links will change once we fully update our Jenkins setup, but
for now, these are the jobs we want to run.