Better error handling for checking if item in s3. PYTHON-2010
Based on feedback and discussion in pull request.

Since `aws s3 ls` does matching by prefix and not on exact file name, a further
check needs to be done before we can say a file exists in s3 or not. To do so,
do a bash string matching.

Separate out checking network connection, AWS keys, and bucket existence into a
separate function. This removes the requirement to check exit codes later on.
More specifically, the `aws s3 ls` command will exit 1 if the file does not
exist, but will exit 255 if there is some sort of connection error. In that
case, we want to continue if the code is 1 and fail if it is 0 or 255.