Refactor upload-to-s3.sh. PYTHON-2010
Drop separate functions that 1) check S3 env settings, and 2) check
if the file exists. That results in 3 calls to AWS. Instead, have a
single function (abort_if_key_exists_or_error), which we can call
one time.

We construct the S3 key with the name of the package. If either the
source tarball or the MD5 file (or both) exist already, running
the `aws s3 ls <KEY_WITH_PACKAGE_NAME>` will return an exit code
of 0, so we can abort. Since we want to abort if either file exists,
there's no need to check explicitly if both keys exist. A single
check will suffice.