Remove packnsend pull_with_restart command. PYTHON-1983
Now that `packnsend start` is more robust and we added a separate
`packnsend restart` command, we can achieve this by combining
existing commands:

    $ packnsend pull
    $ packnsend restart

However, in setup-node.sh, we don't want to restart existing service
containers, even if the docker images have changed, because we don't
want to interfere with other tests that may already be running.

In that case, we want:

    $ packnsend pull
    $ packnsend start

The `packnsend start` command will be a no-op if service containers
are already running, but will start service containers on a new node.

more remove pull_with_restart