Update some test names. PYTHON-1983
This change will help our eyes parse the list of docker tests. Since the
"testPrefix" is only used as the regex with which to sort the tests to their
views, they are best left at the end. This way our eyes will first see the name
of the test, and not a useless prefix.