Support disabling tests. PYTHON-2035
Note that you will get the warning "@Grab support is deprecated
(workspace://test-dsl/jenkins/test-pipeline.groovy, line 1)". More can be found
about this deprecation at
https://github.com/jenkinsci/job-dsl-plugin/wiki/Migration#migrating-to-129.
The way around this, I believe, is to download the required package on our own,
get it into the jenkins workspace, then point to it. However, since Grab still
works, but is just deprecated, I suggest we leave it for now. Grab will no
longer work if we upgrade from jenkins-dsl 1.35 to 1.36, but I'm not sure that
will happen any time soon.

When this becomes an issue (if it does), the required jar file is at
`/var/lib/jenkins/.groovy` on the jaas host. We can always get the jar file
from there, copy it over to the workspace and point to it using the "
Additional classpath" setting.