Add -e to script for Reset Nodes job. PYTHON-1983
If dogestry goes down, then the first command in this file will fail. Without
the -e, the script will continue. Since the Reset Node job is meant to reset
the nodes, it should only succeed if the nodes are actually reset. If a
dogestry pull does not happen, then the nodes cannot be considered reset.