Check if NormalizationRule matches, not if it modifies. PYTHON-2044
Currently, when applying metric grouping rules in the rules engine, a rule has
been considered a "match" if it successfully changed the string. This is
incorrect as per the spec. The rule should be considered a "match" if the regex
returns a match, even if then applying the rule does nothing.

Implement tests for the cross agent fixture that had already been added to the
repo.

Note that the cross agent fixture assumes a case _insensitive_ match whereas we
are currently doing a case _sensitive_ match. A work around for this has been
added to the tests until the issue is addressed.

Adds test environments for python 3.4 and 3.5 for cross agent tests.