Add datastore.instances.r1 feature flag. PYTHON-2075
This feature flag to be used to gate the addition of datastore
instance metrics, as well as instance attributes on transaction
trace segments and slow sql traces.

Because not all database/datastore instrumentation will capture
hostname, port, and database name, we need to be able to set a flag
on the dbapi2 module also. That way we will create new instance
metrics and attributes when both of the following conditions are
true:

1. datastore.instances.r1 is in settings.feature_flag.
2. dbapi2_module._nr_datastore_instance_feature_flag is True.

In our instrumentation, when we call `register_database_client()`, we
will also need to call `enable_datastore_instance_feature()` to set
the feature flag on the dbapi2 module.