Add db instance metrics testing. PYTHON-2065
These tests exercise the database metric reporting capability for
psycopg2. All existing tests are rerun with the datastore.instance.r1
feature_flag added. The expected result is that all psycopg2 tests
should report instance metrics.

In order to ensure all existing tests are run, a new tox file is added
with the feature flag enabled through an ENV variable. At the time
settings are imported, the feature flag is expected to be set (since the
default is loaded from the ENV variable). This behavior allows the
desired instance metric checking to be added at import time.