Add instance metric connect params. PYTHON-2065
Previously, connect parameters were only retained for the purposes of
running explain plans. Now that instance metrics are being reported as
part of the ICAPM updates, connect parameters must be recorded for every
query.

This update ensures connect parameters are passed to the DatabaseTrace
constructor in all recorded database queries for all products.