Add db instance metric reporting. PYTHON-2065
This commit enables datastore instance metric reporting in the
time_metrics method for DatabaseNode.

The instance property for DatabaseNode has been modified to return the
string specified in the agent-specs. Instance will utilize the host/port
to generate the appropriate string.

To enable instance metric reporting, the following must be true:
    - datastore.instance.r1 feature flag is set AND
    - _nr_datastore_instance_feature_flag must be set on the database
      module. This flag is usually set in the database hook by utilizing
      the enable_datastore_instance_feature helper function from within
      the instrument_* procedure.

      Example:
      from newrelic.api.database_trace import enable_datastore_instance_feature
      enable_datastore_instance_feature(module)