Use urllib3 parse_url for pgsql URIs. PYTHON-2101
There is a bug in earlier versions of python in which url parsing
doesn't work for undocumented schemes (such as postgres://). In
addition, ipv6 is not supported by urlparse in python 2.6.

The bugs are documented here:
https://bugs.python.org/issue2987
https://bugs.python.org/issue9374

Since we already vendor the urllib3 package with the agent, we can use
it's parse_url function, rather than rolling our own regex-based
solution.