Wait for postgresql readiness before testing with docker-compose. PYTHON-2103
Adds an entrypoint value to the docker-compose.yml file which will run the
wait_for_ready.sh script before running the given command on the python-base
image. The script uses `psql` to test for the postgresql databases to be ready.

Since `psql` was not already on the python-base image, it was added.