Add default database name for psycopg2. PYTHON-2128
If the database name is not passed in, and the PGDATABASE environment
variable isn't set, then use 'default' as the database name, rather
than None.

Even though postgresql will default to $USER as the database name when
it has not been set explicitly, we would rather not expose the username
automatically, so we've chosen to return 'default' instead.