Teardown module scoped mock calls in data collector. PYTHON-2129
Testing the database_node host resolution requires the ability to create
mock calls or the ability to use the original gethostname from
newrelic.common.

Currently the data_collector monkeypatches the gethostname calls and
does not do any teardown so the behavior of tests become inconsistent
depending on which tests run first.

This commit implements a teardown of the patched components.