Create instance metrics for multipart commands. PYTHON-2143
The Connection object saves info on host/port/db in a _description_args dict.
There are two subclasses of Connection that do the same.

The _instance_info function is split so it only takes the kwargs. This way
these kwargs can be gathered appropriately for the client vs connection
instances.

This commit does not yet send the instance info to the DatastoreTrace.

Several tests were added that show that using _description_args works as we'd
expect it to.