Fix docker/packnsend test image build race condition. PYTHON-2156
With multiple tests running on the same docker host, it was previously
possible to remove the python-base tag while another test is building
FROM python-base. This would result in an attempt to download
python-base from the docker.io registry instead of the internal docker
registry.

The fix in this patch disables the tag removal which eliminates the
race.