Fix alias tests to use Connect() instead of connect(). PYTHON-2154
Also makes the scoped metrics list the same as those found in the
test_cursor.py file.

Prior to this commit, the tests in test_cursor.py and test_alias.py were
identical (aside from the scoped metrics list as noted above). By changing
s/connect()/Connect() we are assuring that we are testing that the Connect()
object is specifically getting wrapped.