Fix Python 3 naming for Flask class based views. PYTHON-2042
The issue at hand is that the `View.as_view` classmethod returns a function
that does not properly set `__qualname__` in Python 3. As per this issue in the
Flask repository https://github.com/pallets/flask/pull/1977.

This commit builds the callable name from the `__module__` and `__name__` attrs
that have been set on the view function, thus removing the need to access
`__qualname__` at all.